% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-0-0plot.R
\name{plot}
\alias{plot}
\title{Plot Cutoff Changes for rdlearn Objects}
\usage{
plot(x, opt, ...)
}
\arguments{
\item{x}{An object of class \code{rdlearn} returned by the
\code{\link{rdlearn}} function.}

\item{opt}{When set to "safe", it displays the derived safe cutoffs and the
original cutoffs. When set to "dif", it displays the change in cutoffs.}

\item{...}{additional arguments.}
}
\value{
A \code{ggplot2} plot which also contains the distance measure between original cutoffs and safe cutoffs.
}
\description{
This function plots the changes in cutoff values relative to the baseline cutoffs
for each group, under different combinations of the smoothness
multiplier (M) and the cost of treatment (C).
}
\examples{
# Simulation Data B from Appendix D of Zhang et al. (2022)
set.seed(1)
n <- 300
X <- runif(n, -1000, -1)
G <- 2 * as.numeric(
I(0.01 * X + 5 + rnorm(n, sd = 10) > 0)
) +
as.numeric(
I(0.01 * X + 5 + rnorm(n, sd = 10) <= 0)
)
c1 <- -850
c0 <- -571
C <- ifelse(G == 1, c1, c0)
D <- as.numeric(X >= C)
coef0 <- c(-1.992230e+00, -1.004582e-02, -1.203897e-05, -4.587072e-09)
coef1 <- c(9.584361e-01, 5.308251e-04, 1.103375e-06, 1.146033e-09)
Px <- poly(X, degree = 3, raw = TRUE)
# Px = poly(X-735.4334-c1,degree=3,raw=TRUE) for Simulation A
Px <- cbind(rep(1, nrow(Px)), Px)
EY0 <- Px \%*\% coef0
EY1 <- Px \%*\% coef1
d <- 0.2 + exp(0.01 * X) * (1 - G) + 0.3 * (1 - D)
Y <- EY0 * (1 - D) + EY1 * D - d * as.numeric(I(G == 1)) + rnorm(n, sd = 0.3)

simdata_B_demo <- data.frame(Y,X,C)

# Learn new treatment assignment cutoffs
rdlearn_result <- rdlearn(
  y = "Y", x = "X", c = "C", data = simdata_B_demo,
  fold = 2, M = 0, cost = 0
)

# Summarise the learned policies
summary(rdlearn_result)

# Visualize the learned policies
plot(rdlearn_result, opt = "dif")
# The learned cutoff for Group 1 is the same as the baseline cutoff, because
# the baseline cutoff is set to equal to oracle cutoff in this simulation.

# Implement sensitivity analysis
sens_result <- sens(rdlearn_result, M = 1, cost = 0)
plot(sens_result, opt = "dif")
}
