% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_0plot.rdlearn.R
\name{plot.rdlearn}
\alias{plot.rdlearn}
\title{Plot method for \code{rdlearn} objects}
\usage{
\method{plot}{rdlearn}(result, xlab = "", ylab = "")
}
\arguments{
\item{result}{an object of class \code{rdlearn} returned by the \code{\link{rdlearn}}.}

\item{xlab}{a label of x-axis.}

\item{ylab}{a label of y-axis.}
}
\value{
a \code{ggplot2} plot of changes in cutoffs.
}
\description{
\code{plot} plots the cutoff change relative to the baseline cutoff for each department (y-axis)
under different smoothness multiplicative factor M and cost of treatment C (x-axis).
}
