% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdlearn_package.R
\docType{package}
\name{RDDPackage-package}
\alias{RDDPackage}
\alias{RDDPackage-package}
\title{rdlearn: A package for...}
\description{
The \code{rdlearn} package provides three categories of important functions:
rdlearn, plot and sens.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kkawato/RDDPackage}
  \item Report bugs at \url{https://github.com/kkawato/RDDPackage/issues}
}

}
