% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_0rdlearn.R
\name{rdlearn}
\alias{rdlearn}
\title{Safe policy learning under regression discontinuity designs with multiple cutoffs}
\usage{
rdlearn(y, x, c, groupname = NULL, data, fold = 10, M = 1, cost = 0)
}
\arguments{
\item{y}{column name of outcome variable.}

\item{x}{column name of running variable.}

\item{c}{column name of cutoff.}

\item{groupname}{column name of each cutoff group's name (e.g. department name).
If no argument is entered, the names "Group 1", "Group 2", ... are
assigned from the group with smallest cutoff.}

\item{data}{data frame containing all variables.}

\item{fold}{number of folds for cross-fitting. Default is 10.}

\item{M}{multiplicative smoothness factor. Default is 1.}

\item{cost}{cost for calculating regret.
Default is 0. Cost should be scaled by the range of the outcome Y.}
}
\value{
an \code{rdlearn} object containing ...
}
\description{
\code{rdlearn} estimates
}
\references{
Yi Zhang ...
}
