% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-0-0simdata_B.R
\docType{data}
\name{simdata_B}
\alias{simdata_B}
\title{Simulation Data B}
\format{
A data frame with 2000 rows and 3 columns:
\describe{
  \item{Y}{outcome variable}
  \item{X}{running variable}
  \item{C}{cutoff value}
  }
}
\usage{
simdata_B
}
\description{
This dataset is based on the ACCES Program and generated according to
scenario B described in Appendix D of Zhang et al. (2022). In this scenario,
the baseline policy (cutoff) differs from the oracle policy.
}
\references{
Zhang, Y., Ben-Michael, E. and Imai, K. (2022) 'Safe Policy
  Learning under Regression Discontinuity Designs with Multiple Cutoffs',
  arXiv [stat.ME]. Available at: \url{http://arxiv.org/abs/2208.13323}.
}
\keyword{datasets}
