% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-0-0summary.R
\name{summary}
\alias{summary}
\title{Summary function}
\usage{
summary(result)
}
\arguments{
\item{result}{An object of class \code{rdlearn} returned by the
\code{\link{rdlearn}} function.}
}
\value{
Displays key outputs from the \code{\link{rdlearn}} function. It
  provides basic information and RD causal effect estimates from
  \code{\link{rdesimate}}, as well as the safe cutoffs derived by
  \code{\link{rdlearn}} and the difference between them and the original
  cutoffs.
}
\description{
Summary function
}
