% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-0-0summary.R
\name{summary.rdlearn}
\alias{summary.rdlearn}
\title{Summary function}
\usage{
\method{summary}{rdlearn}(object, ...)
}
\arguments{
\item{object}{An object of class \code{rdlearn} returned by the
\code{\link{rdlearn}} function.}

\item{...}{additional arguments.}
}
\value{
Displays key outputs from the \code{\link{rdlearn}} function. It
  provides basic information and RD causal effect estimates from
  \code{\link{rdestimate}}, as well as the safe cutoffs derived by
  \code{\link{rdlearn}} and the difference between them and the original
  cutoffs.
}
\description{
Summary function
}
\examples{
data(acces)
rdlearn_result <- rdlearn(
  y = "elig", x = "saber11", c = "cutoff",
  group_name = "department", data = acces,
  fold = 5, M = c(0, 1), cost = 0
)
summary(rdlearn_result)
plot(rdlearn_result, opt = "dif")

}
