% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-5-0calculate_distance.R
\name{calculate_distance}
\alias{calculate_distance}
\title{Calculate distance measure between original cutoffs and safe cutoffs}
\usage{
calculate_distance(org_cut, safe_cut)
}
\arguments{
\item{org_cut}{Numeric vector of original cutoffs.}

\item{safe_cut}{Data frame of safe cutoffs. The first column should be the
group column, and subsequent columns the cutoffs.}
}
\value{
A numeric vector containing the distance between safe cutoffs and
  original cutoffs, measured in terms of the L1 norm, L2 norm, and uniform
  norm.
}
\description{
Calculate distance measure between original cutoffs and safe cutoffs
}
