% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_0plot.rdlearn.R
\name{plot.rdlearn}
\alias{plot.rdlearn}
\title{Plot Cutoff Changes for rdlearn Objects}
\usage{
\method{plot}{rdlearn}(result, xlab = "", ylab = "")
}
\arguments{
\item{result}{An object of class \code{rdlearn} returned by the \code{\link{rdlearn}} function.}

\item{xlab}{A character string specifying the label for the x-axis.}

\item{ylab}{A character string specifying the label for the y-axis.}
}
\value{
A \code{ggplot2} object representing the plot of cutoff changes.
}
\description{
This function plots the changes in cutoff values relative to the baseline cutoffs
for each group, under different combinations of the smoothness
multiplier (M) and the cost of treatment (C).
}
\examples{
result <- rdlearn(y = "elig", x = "saber11", c = "cutoff",
                  group_name = "department", data = acces,
                  fold = 20, M = c(0, 1), cost = 0)
plot(result)

}
