% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-0-0plot.R
\name{plot}
\alias{plot}
\title{Plot Cutoff Changes for rdlearn Objects}
\usage{
plot(result, opt)
}
\arguments{
\item{result}{An object of class \code{rdlearn} returned by the \code{\link{rdlearn}} function.}

\item{opt}{"safe"と指定することによって、derived safe cutoffsとoriginal cutoffsを表示する。
"dif"と指定することによって、change in cutoffsを表示する。}
}
\value{
A \code{ggplot2} plot which also contains the.
}
\description{
This function plots the changes in cutoff values relative to the baseline cutoffs
for each group, under different combinations of the smoothness
multiplier (M) and the cost of treatment (C).
}
\examples{
\dontrun{
# Load example data
data(acces)
library(rdlearn)
library(nprobust)
library(nnet)
library(ggplot2)
library(dplyr)
library(glue)
library(purrr)
library(tidyr)

result <- rdlearn(y = "elig", x = "saber11", c = "cutoff",
                  group_name = "department", data = acces,
                  fold = 20, M = c(0, 1), cost = 0)
plot(result)
}

}
