% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-0-0plot.R
\name{plot.rdlearn}
\alias{plot.rdlearn}
\title{Plot Cutoff Changes for rdlearn Objects}
\usage{
\method{plot}{rdlearn}(x, opt, ...)
}
\arguments{
\item{x}{An object of class \code{rdlearn} returned by the
\code{\link{rdlearn}} function.}

\item{opt}{When set to "safe", it displays the derived safe cutoffs and the
original cutoffs. When set to "dif", it displays the change in cutoffs.}

\item{...}{additional arguments.}
}
\value{
A \code{ggplot2} plot which also contains the distance measure
  between original cutoffs and safe cutoffs.
}
\description{
This function plots the changes in cutoff values relative to the baseline
cutoffs for each group, under different combinations of the smoothness
multiplier (M) and the cost of treatment (C).
}
\examples{
# Load the simulation data
data(simdata_A)

# Learn new treatment assignment cutoffs
rdlearn_result <- rdlearn(
  y = "Y", x = "X", c = "C", data = simdata_A,
  fold = 2, M = c(0, 1), cost = 0
)

# Summarise the learned policies
summary(rdlearn_result)

# Visualize the learned policies
plot(rdlearn_result, opt = "dif")

# Implement sensitivity analysis
sens_result <- sens(rdlearn_result, M = 1, cost = c(0,1))
plot(sens_result, opt = "dif")
}
