% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-0-0plot.R
\name{plot}
\alias{plot}
\title{Plot Cutoff Changes for rdlearn Objects}
\usage{
plot(result, xlab = "", ylab = "")
}
\arguments{
\item{result}{An object of class \code{rdlearn} returned by the \code{\link{rdlearn}} function.}

\item{xlab}{A character string specifying the label for the x-axis.}

\item{ylab}{A character string specifying the label for the y-axis.}
}
\value{
A \code{ggplot2} object representing the plot of cutoff changes.
}
\description{
This function plots the changes in cutoff values relative to the baseline cutoffs
for each group, under different combinations of the smoothness
multiplier (M) and the cost of treatment (C).
}
