% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-0rdlearn_package.R
\docType{package}
\name{RDDPackage-package}
\alias{RDDPackage}
\alias{RDDPackage-package}
\title{rdlearn: A package for}
\description{
The \code{rdlearn} package provides three categories of important functions:
rdlearn, plot and sens.
}
\references{
Zhang, Yi, Eli Ben-Michael, and Kosuke Imai. 2023. “Safe Policy
 Learning under Regression Discontinuity Designs with Multiple Cutoffs.”
 arXiv [stat.ME]. arXiv. http://arxiv.org/abs/2208.13323.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kkawato/RDDPackage}
  \item Report bugs at \url{https://github.com/kkawato/RDDPackage/issues}
}

}
