% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-0-0rdlearn.R
\name{rdlearn}
\alias{rdlearn}
\title{Safe Policy Learning for Regression Discontinuity Design with Multiple
Cutoffs}
\usage{
rdlearn(
  y,
  x,
  c,
  group_name = NULL,
  data,
  fold = 10,
  M = 1,
  cost = 0,
  trace = TRUE
)
}
\arguments{
\item{y}{A character string specifying the name of the outcome variable
column.}

\item{x}{A character string specifying the name of the running variable
column.}

\item{c}{A character string specifying the name of the cutoff variable
column.}

\item{group_name}{A character string specifying the name of the column
containing group names (e.g., department names) for each cutoff. If not
provided, the groups are assigned names "Group 1", "Group 2", ... in
ascending order of cutoff values.}

\item{data}{A data frame containing all required variables.}

\item{fold}{The number of folds for cross-fitting. Default is 10.}

\item{M}{A numeric value or vector specifying the multiplicative smoothness
factor(s) for sensitivity analysis. Default is 1.}

\item{cost}{A numeric value or vector specifying the cost of treatment for
calculating regret. This cost should be scaled by the range of the outcome
variable Y. Default is 0.}

\item{trace}{A logical value that controls whether to display the progress.
If set to TRUE, the progress will be printed. The default value is TRUE.}
}
\value{
An object of class \code{rdlearn}, which is a list containing the
  following components:
  \describe{
    \item{call}{The original function call.}
    \item{var_names}{A list of variable names for the outcome, running variable, and cutoff.}
    \item{org_cut}{A vector of original cutoff values.}
    \item{safe_cut}{A data frame containing the obtained new treatment assignment cutoffs.}
    \item{sample}{The total sample size.}
    \item{num_group}{The number of groups.}
    \item{group_name}{A vector of group names.}
    \item{cross_fit_output}{The intermediate output of the cross-fitting procedure.}
    \item{dif_lip_output}{The intermediate output of the cross-group differences and the smootheness parameters}
  }
}
\description{
The \code{rdlearn} function implements safe policy learning under a
regression discontinuity design with multiple cutoffs. The resulting new
treatment assignment rules (cutoffs) are guaranteed to yield no worse overall
outcomes than the existing cutoffs.
}
\examples{
\dontrun{
# Load example data
data(acces)
library(rdlearn)
library(nprobust)
library(nnet)
library(ggplot2)
library(dplyr)
library(glue)
library(purrr)
library(tidyr)

result <- rdlearn(y = "elig", x = "saber11", c = "cutoff",
                  group_name = "department", data = acces,
                  fold = 20, M = c(0, 1), cost = 0)
plot(result)
}

}
