% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-0-0sens.R
\name{sens}
\alias{sens}
\title{Sensitivity Analysis for rdlearn Objects}
\usage{
sens(object, M = NULL, cost = NULL, trace = TRUE)
}
\arguments{
\item{object}{An object of class \code{rdlearn} returned by the
\code{\link{rdlearn}} function.}

\item{M}{A numeric value or vector specifying the multiplicative smoothness
factor(s) for sensitivity analysis.}

\item{cost}{A numeric value or vector specifying the cost of treatment for
calculating regret.}

\item{trace}{A logical value that controls whether to display the progress of
cross-fitting and regret calculation. If set to TRUE, the progress will be
printed. The default value is TRUE.}
}
\value{
An updated \code{rdlearn} object with the new cutoffs based on the
  provided values of M and cost.
}
\description{
This function performs sensitivity analysis for the \code{rdlearn} object
under different smoothness multiplier (M) and the cost of treatment (cost).
}
\examples{
data(acces)

# Learn new treatment assignment cutoffs
rdlearn_result <- rdlearn(
  y = "elig", x = "saber11", c = "cutoff",
  group_name = "department", data = acces,
  fold = 5, M = c(0, 1), cost = 0
)

# Summarise the learned policies
summary(rdlearn_result)
# Visualize the learned policies
plot(rdlearn_result, opt = "dif")

# Implement sensitivity analysis
sens_result <- sens(rdlearn_result, M = 1, cost = c(0,1))
plot(sens_result, opt = "dif")
}
