% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-0-0sens.R
\name{sens}
\alias{sens}
\title{Sensitivity Analysis for rdlearn Objects}
\usage{
sens(object, M = NULL, cost = NULL, trace = TRUE)
}
\arguments{
\item{object}{An object of class \code{rdlearn} returned by the
\code{\link{rdlearn}} function.}

\item{M}{A numeric value or vector specifying the multiplicative smoothness
factor(s) for sensitivity analysis.}

\item{cost}{A numeric value or vector specifying the cost of treatment for
calculating regret.}

\item{trace}{A logical value that controls whether to display the progress of
cross-fitting and regret calculation. If set to TRUE, the progress will be
printed. The default value is TRUE.}
}
\value{
An updated \code{rdlearn} object with the new cutoffs based on the
  provided values of M and cost.
}
\description{
This function performs sensitivity analysis for the \code{rdlearn} object
under different smoothness multiplier (M) and the cost of treatment (cost).
}
\examples{
\dontrun{
# Load example data
data(acces)
library(rdlearn)
library(nprobust)
library(nnet)
library(ggplot2)
library(dplyr)
library(glue)
library(purrr)
library(tidyr)
result <- rdlearn(
  y = "elig", x = "saber11", c = "cutoff",
  group_name = "department", data = acces,
  fold = 20, M = c(0, 1), cost = 0
)
sens_result <- sens(result, M = 1, cost = c(0, 0.2, 0.4, 0.6, 0.8, 1))
plot(sens_result)
}
}
