% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_0sens.rdlearn.R
\name{sens.rdlearn}
\alias{sens.rdlearn}
\title{Sensitivity analysis method for \code{rdlearn} objects}
\usage{
\method{sens}{rdlearn}(result, M = NULL, cost = NULL)
}
\arguments{
\item{result}{An object of class \code{rdlearn} returned by the \code{\link{rdlearn}}.}

\item{M}{A multiplicative smoothness factor.}

\item{cost}{A cost for calculating regret.}
}
\value{
an \code{rdlearn} object containing ...
}
\description{
\code{sens}
}
\examples{
result2 <- sens(result, M = 1, cost=c(0, 0.2, 0.4, 0.6, 0.8, 1))
plot(result2)

}
