% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_0sens.rdlearn.R
\name{sens.rdlearn}
\alias{sens.rdlearn}
\title{Sensitivity analysis method for \code{rdlearn} objects}
\usage{
\method{sens}{rdlearn}(result, M = NULL, cost = NULL)
}
\arguments{
\item{result}{an object of class \code{rdlearn} returned by the \code{\link{rdlearn}}.}

\item{M}{multiplicative smoothness factor.}

\item{cost}{cost for calculating regret.}
}
\value{
an \code{rdlearn} object containing ...
}
\description{
\code{sens}
}
